+++ src/ShapeCrawler/Extensions/TypedOpenXmlPartExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using System.IO;
 using DocumentFormat.OpenXml.Packaging;
-using ShapeCrawler.Shared;
+using ShapeCrawler.Presentations;
 
 namespace ShapeCrawler.Extensions;
