+++ src/ShapeCrawler/Fonts/TextPortionFont.cs
@@ -16,7 +16,7 @@ internal sealed class TextPortionFont : ITextPortionFont
     private readonly Lazy<FontColor> fontColor;
     private readonly IFontSize size;
     private readonly ThemeFontScheme themeFontScheme;
-    private readonly SCAText sdkScaText;
+    private readonly SCAText scAText;
 
     internal TextPortionFont(
         OpenXmlPart sdkTypedOpenXmlPart,
@@ -41,7 +41,7 @@ internal sealed class TextPortionFont : ITextPortionFont
         this.fontColor = new Lazy<FontColor>(() => new FontColor(sdkTypedOpenXmlPart, this.aText));
         this.size = size;
         this.themeFontScheme = themeFontScheme;
-        this.sdkScaText = new SCAText(sdkTypedOpenXmlPart, aText);
+        this.scAText = new SCAText(sdkTypedOpenXmlPart, aText);
     }
 
     #region Public APIs
@@ -68,8 +68,8 @@ internal sealed class TextPortionFont : ITextPortionFont
 
     public string EastAsianName
     {
-        get => this.sdkScaText.EastAsianName();
-        set => this.sdkScaText.UpdateEastAsianName(value);
+        get => this.scAText.EastAsianName();
+        set => this.scAText.UpdateEastAsianName(value);
     }
 
     public bool IsBold
