+++ src/ShapeCrawler/GroupShapes/GroupedShape.cs
@@ -1,7 +1,7 @@
 ﻿using System.Linq;
 using DocumentFormat.OpenXml;
 using ShapeCrawler.ShapeCollection;
-using ShapeCrawler.Shared;
+using ShapeCrawler.Units;
 using P = DocumentFormat.OpenXml.Presentation;
 
 namespace ShapeCrawler.GroupShapes;
