+++ src/ShapeCrawler/Presentations/SlideSize.cs
@@ -1,5 +1,5 @@
 ﻿using DocumentFormat.OpenXml;
-using ShapeCrawler.Shared;
+using ShapeCrawler.Units;
 using P = DocumentFormat.OpenXml.Presentation;
 
 namespace ShapeCrawler.Presentations;
