+++ src/ShapeCrawler/ShapeCollection/SCSlidePart.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using DocumentFormat.OpenXml;
-using ShapeCrawler.Shared;
+using ShapeCrawler.Presentations;
 using ShapeCrawler.Units;
 
 namespace ShapeCrawler.ShapeCollection;
