+++ src/ShapeCrawler/ShapeCollection/SlideShapeOutline.cs
@@ -2,7 +2,7 @@
 using DocumentFormat.OpenXml.Packaging;
 using ShapeCrawler.Drawing;
 using ShapeCrawler.Extensions;
-using ShapeCrawler.Shared;
+using ShapeCrawler.Units;
 using A = DocumentFormat.OpenXml.Drawing;
 
 namespace ShapeCrawler.ShapeCollection;
